Issue #4: Removed the routes file and moved all the functions into the server.js file.
